+++ executorCommon/src/main/scala/dx/executor/JobMeta.scala
@@ -21,6 +21,7 @@ import dx.api.{
 }
 import dx.core.Constants
 import dx.core.io.DxWorkerPaths
+import dx.core.ir.RunSpec.InstanceType
 import dx.core.ir.Value.VNull
 import dx.core.ir.{
   DxName,
@@ -727,12 +728,21 @@ abstract class JobMeta(val workerPaths: DxWorkerPaths,
 
   lazy val parserOptions: Option[JsValue] = getExecutableDetail(Constants.ParseOptions)
 
-  lazy val instanceTypeDb: InstanceTypeDB = getExecutableDetail(Constants.InstanceTypeDb) match {
-    case Some(JsString(s)) =>
-      val js = CodecUtils.base64DecodeAndGunzip(s)
-      js.parseJson.convertTo[InstanceTypeDB]
-    case other =>
-      throw new Exception(s"unexpected ${Constants.InstanceTypeDb} value ${other}")
+  lazy val instanceTypeDb: InstanceTypeDB = {
+    try {
+      // first try to create the db from the current project
+      InstanceType.createDb(dxApi = dxApi)
+    } catch {
+      case _: Throwable =>
+        // fall back to the cached database if available
+        getExecutableDetail(Constants.InstanceTypeDb) match {
+          case Some(JsString(s)) =>
+            val js = CodecUtils.base64DecodeAndGunzip(s)
+            js.parseJson.convertTo[InstanceTypeDB]
+          case other =>
+            throw new Exception(s"unexpected ${Constants.InstanceTypeDb} value ${other}")
+        }
+    }
   }
 
   lazy val defaultRuntimeAttrs: Map[String, Value] =