+++ executorWdl/src/test/scala/dx/executor/wdl/WdlTaskExecutorTest.scala
@@ -151,10 +151,10 @@ class TaskExecutorTest extends AnyFlatSpec with Matchers {
                                Vector(Constants.OsVersion))
       ),
       Some(DiskType.SSD),
-      Some(1.00f)
+      Some(1)
   )
   private val instanceTypeDB =
-    InstanceTypeDB(Map(TaskTestJobMeta.InstanceType -> unicornInstance), pricingAvailable = true)
+    InstanceTypeDB(Map(TaskTestJobMeta.InstanceType -> unicornInstance))
 
   private def pathFromBasename(basename: String): Option[Path] = {
     getClass.getResource(s"/task_runner/${basename}") match {